Refactor import statements to use the correct Streak model in profile module
