+++ src/app/models/Streak.ts
