+++ src/app/models/types/responses/listStreaksResponse.interface.ts
@@ -1,4 +1,4 @@
-import { Streak } from '../../Streaks';
+import { Streak } from '../../Streak';
 
 export interface listStreaksResponseInterface {
   total: number;
